requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
Levenshtein==0.25.1
multidict==6.0.5
numpy==2.0.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
rapidfuzz==3.9.4
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
